requirements.txt
fastapi 
uvicorn
jupyterlab
openai
python-dotenv